+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 - Fix deprecated CircleCI config [#3181](https://github.com/opendatateam/udata/pull/3181)
 - Use proper RESTful Hydra API endpoints [#3178](https://github.com/opendatateam/udata/pull/3178)
 - Add a "filter by organization badge" for datasets, dataservices, reuses and organizations [#3155](https://github.com/opendatateam/udata/pull/3155]
+- Add dataservices search with search-service [#3171](https://github.com/opendatateam/udata/pull/3171)
+    * you will need https://github.com/opendatateam/udata-search-service/pull/48
 
 ## 9.2.4 (2024-10-22)